add rename attribute

+++ erdos/graph.py
@@ -197,12 +197,14 @@ class Graph(object):
                     stream_names.add(s.name)
                 else:
                     old_name = s.name
-                    tf = time.time() * 1000
-                    s.rename(tf)
+                    tf = int(time.time() * 1000)
+                    s.rename(tf)  # attach 13 digits to the original name
                     for dependant_id in op_handle.dependant_ops:
                         for input in self.op_handles[dependant_id].input_streams:
                             if input.name == old_name:
                                 input.rename(tf)
+
+        # Build sink graph
         self._build_output_stream_sinks_graph()
 
     def _build_output_stream_sinks_graph(self):
